requirements.txt
pillow
pillow-avif-plugin